pyproject.toml
[project]
name = "aws-cdk-agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aws-lambda-powertools>=3.24.0",
    "aws-opentelemetry-distro>=0.15.0",
    "bedrock-agentcore>=1.3.1",
    "fastapi>=0.129.2",
    "feedparser>=6.0.12",
    "nanoid>=2.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.1",
    "pynamodb>=6.1.0",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "strands-agents>=1.27.0",
    "strands-agents-tools>=0.2.21",
    "tavily>=1.1.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.1",
]

[tool.ruff]
target-version = "py314"

lint.ignore = [
    "E501",  # line too long
]

lint.select = [
    "F",  # pyflakes
    "E",  # pycodestyle
    "W",  # pycodestyle warnings
    "I",  # isort
    "UP", # pyupgrade
    "N",  # pep8-naming
]
